Add files via upload
first electronics project. I wish to make a safe box powered and controlled by a pi. I will first begin with 2 sage authentication but will hopefully make it 3 if all goes well. This is just initial brain storming and outlining edge cases that need to be addressed. I loved the servo so much that I will use it as the main locking mechanism rather than my initial thought of an electro-magnet that would use far too much power and effort.